Add lib.misc.composable
